TASK #4355 LTE网格数据管理功能增强

+++ src/main/java/com/hgicreate/rno/domain/OriginFileAttr.java
@@ -15,7 +15,7 @@ public class OriginFileAttr {
     @SequenceGenerator(name = "OriginFileAttrSeq", sequenceName = "SEQ_ORIGIN_FILE_ATTR", allocationSize = 1)
     private Long id;
 
-    private Integer originFileId;
+    private Long originFileId;
     private String name;
     private String value;
